// src/clients/scaffoldingService/models/ScaffoldV1TemplateSpec.ts
export interface ScaffoldV1TemplateSpec {
  /**
   * The name of the template.
   * @type {string}
   * @memberof ScaffoldV1TemplateSpec
   */
  name?: string;
  /**
   * The display name of the template.
   * @type {string}
   * @memberof ScaffoldV1TemplateSpec
   */
  display_name?: string;
  /**
   * The description of the template.
   * @type {string}
   * @memberof ScaffoldV1TemplateSpec
   */
  description?: string;
  /**
   * The semantic version of the template, of the form <major>.<minor>.<patch>. Bump the version with each non-trivial change.
   * @type {string}
   * @memberof ScaffoldV1TemplateSpec
   */
  version?: string;
  /**
   * The primary programming language that projects generated from the template will use.
   * @type {string}
   * @memberof ScaffoldV1TemplateSpec
   */
  language?: string;
  /**
   * Tags are keywords or very short phrases (1-3 words) that apply to the template and that help with search and organization. Use lowercase characters and numbers in each word, and separate words with a space.
   * @type {Array<string>}
   * @memberof ScaffoldV1TemplateSpec
   */
  tags?: Array<string>;
  /**
   * Options of the template.
   * @type {{ [key: string]: ScaffoldV1TemplateOption; }}
   * @memberof ScaffoldV1TemplateSpec
   */
  options?: { [key: string]: ScaffoldV1TemplateOption };
  /**
   * The template_collection to which this belongs.
   * @type {GlobalObjectReference}
   * @memberof ScaffoldV1TemplateSpec
   */
  template_collection?: GlobalObjectReference;
}